simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.301",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.1.0",
]
